refactor: simplify streaming example
