+++ guides/streaming.md
@@ -1,9 +1,5 @@
 # Streaming Markdown
 
-> #### Unreleased {: .warning}
->
-> This API is on the development branch. It is not on Hex yet.
-
 Use `MDEx.stream/2` when Markdown arrives in chunks from an LLM, file, or HTTP
 response.
 
@@ -198,19 +194,17 @@ bytes. An incomplete code point at EOF also raises.
 
 ## Req response streams
 
-Req can feed a Stream from its callback API. This keeps backpressure when the
-consumer deliberately slows down:
+Req's asynchronous response body is an Enumerable, so it can feed MDEx
+directly:
 
 ```elixir
-url
-|> req_stream()
+Req.get!(url, into: :self).body
 |> MDEx.stream(extension: [table: true])
 |> Enum.each(&consume/1)
 ```
 
-The body does not need to be joined first. The Phoenix example includes a
-complete `MDExStreamingDemo.HTTPStream` adapter with a body-size limit,
-redirect checks, public URL validation, and cancellation.
+The body does not need to be joined first. Req cancels the asynchronous request
+if the consumer stops early.
 
 ## Partial Markdown
 
@@ -354,7 +348,7 @@ is unique only within one Stream run.
 runs this path:
 
 ```text
-Req callback-backed Stream
+Req.Response.Async Enumerable
   -> MDEx.stream/2
   -> one {id, document} chunk per message
   -> Phoenix.LiveView.stream_insert/4
@@ -367,11 +361,9 @@ Run it from the repository:
 elixir examples/streaming.exs
 ```
 
-The URL defaults to the raw MDEx README on GitHub. The example accepts public
-HTTP(S) URLs, validates every redirect, and enforces the body-size limit before
-chunks enter the paced pipeline. It splits large network reads into 64-byte
-chunks so updates stay visible. The controls include a delay of up to 1000 ms
-and optional auto-scroll.
+The URL defaults to the raw MDEx README on GitHub. The example splits large
+network reads into 64-byte chunks so updates stay visible. The controls include
+a delay of up to 1000 ms and optional auto-scroll.
 
 Lumis highlights partial code fences. The page also shows source counts, MDEx
 updates, DOM chunks, and current and peak memory for the producer and LiveView
